Updated mtsLogMessage class
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@3241 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstMultiTask/code/mtsParameterTypes.cpp
@@ -417,7 +417,7 @@ mtsLogMessage::mtsLogMessage(const mtsLogMessage & other)
     memcpy(this->Message, other.Message, this->Length);
 }
 
-mtsLogMessage::mtsLogMessage(const char * log, size_t len) 
+mtsLogMessage::mtsLogMessage(const char * log, unsigned int len) 
     : mtsGenericObject(),
       Length(len), ProcessName("")
 {
